Add the missing evemu_set_name() function
During the evolution of evemu, the process of setting the device
name has shifted towards the evemu_extract() and evemu_read() functions,
rendering the initial name in evemu_new() more or less useless.
This patch adds the missing evemu_set_name() function, in order to
make it possible to override the device name before device creation.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>