Add evemu-event to play a single event from the commandline
Some testing scenarios require specific events to be generated that are hard
to create from a physical device (e.g. touchpad motion purely on the
positive X axis with no Y axis). For those scenarios, recording a device and
hacking the evemu event files is time consuming.

evemu-event simply allows this scenario to be scripted.