Write event descriptions as one block up to EV_SYN
Instead of writing the events out as they come in, write them out as one
continuous block up to EV_SYN. This makes comments a lot more readable, and
newer kernels won't send the events until EV_SYN is received anyway.

assert() to handle failure conditions is enough here. This is very unlikely
to ever get triggered and the code to handle this would be essentially
unmaintained.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>