tools: drop evemu-describe, merge into evemu-record
evemu-record will print a device description first before the actual event
sequence. evemu-describe is the same binary, it'll just quit after printing
the description.

Now that evemu-record contains the description, evemu-play will leave when
hitting this description. Thus the change in evemu.c to make it ignore
any lines not starting with "E:".

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>