Warn about replaying incompatible events
Shout a warning if an event is about to be replayed that's not actually
supported by the device the fd points to. since evemu_play() only has a FILE
and fd API we need to create a device on-the-fly. If that fails, we pretend
everything is ok and we just won't be able to warn.

Evemu Bug 63061 <http://bugs.freedesktop.org/show_bug.cgi?id=63061>

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>