Un-constify evemu_create()'s evemu_device struct argument
Having the 'create' API take a const argument is a really bad idea, it makes
it impossible to store any data inside the struct that would be useful on
creation. This includes automatically setting the node name on creation, or,
more importantly here, enabling a new libevdev backend.

Since this is an ABI break, bump the libtool version.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>