py: Don't use find_library for libevemu.so
The current mechanism to find libevemu.so prefered a system wide
installed version and if found it ignored the in tree build one.
That's bad for running the tests as they should be performed on the in
tree build library.

So, don't use find_library for libevemu.so and let CDLL do its job. It
even takes LD_LIBRARY_PATH into account, which find_library doesn't.

Signed-off-by: Daniel Martin <consume.noise@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>