py: Fix file object detection
Python 3 uses different classes for opened text and binary files.
Therefore a test on __name__ would need to cover more (i.e.
BufferedReader and TextIOWrapper). But, we just want to read from the
file objects and a test for the attribute "read" is sufficient and even
more generic.

Additionally, just check if the type of an object is str and not if its
__name__ is "str".

Signed-off-by: Daniel Martin <consume.noise@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>