py: Generate complete dicts in event_names.py
Before, the generated dictionaries in event_names.py just had the bits
to name mapping and a loop to add the name to bits mapping using items()
afterwards. As of Python 3 items() returns a view, not a copy, resulting
in an error as the dictionary was modified in place.

Now, generate both the bits to name and name to bits mapping when
creating event_names.py and make the loop obsolete.

Signed-off-by: Daniel Martin <consume.noise@gmail.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>