py: Add base class LibraryWrapper
LibraryWrapper will be a base class for others wrapping a shared
library with ctypes.

And add various functions (i.e. expect_ge_zero()), which will be used as
callback functions to check the return value of an API call.

Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Daniel Martin <consume.noise@gmail.com>

Squashed: py: Add ENCODING="iso8859-1" to const.py
The ENCODING is used to decode bytes to str objects

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>