Call write_event_desc from evemu_write_event
This was exported API and we're now changing the behavior here:
- adding a human-readable comment for each event. technically the output of
  evemu_write_event isn't compatible with evemu 1.0 output anymore now
- a newline character is now appended to the output

I don't expect either to be an issue though.

The return value of evemu_write_event is just documented as "Returns a
positive number if successful, zero or negative error otherwise." but just in
case someone was using the return value as the length of the printed string
keep that behavior.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>