python: drop evemu.const usages for device bits
We have make-event-names.py that generates the required mappings, there is no
need for having a hand-updated file around that provides some of these
constants.

Previously we compared the event names as well as the event values in
test_case.py. That was useful when we were using two different sources
(event_names.py and the hand-maintained const.py). Now we generate the dicts
from the same source. We have separate tests for evemu.event_get_(value|name)
so we don't need the this extra comparison against the names in test_get_abs_*
and friends anymore.

Special note on get_expected_propbits(): we previously returned a dict with
the absolute axes as keys. This is fixed here, we now return a dict with the
actual property bits as keys. Since all values were always False and both
absbits and propbits are just numbers this never got noticed before.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>