python: switch from type() == str to not isinstance(int)
Switch to isinstance to be more liberal in what we can accept.
Anything that looks like an int should be used as such, anything else we
assume to be a string that needs parsing. event_get_value() would do that for
us but we can't call it unconditionally - some events don't have names so they
cannot resolve.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>