Record the state of SW/LED in the evemu file
There is no evdev ioctl to set the state of a device without sending an event
through the device, so we can't replicate the state without altering the
recorded event sequence. For this reason, we use the libevdev shortcut to
query the state rather than adding an evemu-specific API to get the state.

The state is printed to the comments, and optionally printed to the recording
(0 is assumed when missing).

EV_KEY can give us the state but we don't print it, unlike the others a key
state is too transient.
EV_SND could be added, if we ever see demand for it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>