Record events with an initial offset of 1µs
Since 63df84261355d, the first event frame always has a timestamp of 0.0. When
reading those events we'll replay the first event frame, and set evtime. On
the second frame, evtime is 0.0 which evemu_read_event_realtime() takes as
"uninitialized" and thus resets it again, causing the second frame to be
replayed immediately instead of waiting for the correct time.

Since this behavior is part of the API, work around this by giving the first
event an offset of 1µs. Since the 0 offset was to make reading event logs
easier, this offset won't hurt us. It won't change existing recordings, but
over time this will be less of an issue and it's less of a hack than trying to
work magic values into evemu_play/evemu_read_event_realtime.

The alternative to this hack would be to add another API (or break the current
one), but it's not yet worth it for the little gain we get from it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>