evemu-record: add support for --autorestart to cycle buffers
Uses the existing evemu timeout feature where evemu_record() quits after a
given timeout. A call of

   evemu-record --autorestart 10 /dev/input/event4 scroll.evemu

will use the filename as prefix and create a timestamped output file (e.g.
scroll.evemu.2016-02-25-09:13). Until the 10s inactivity timeout is hit, all
events are recorded to that file. After that, the file is closed with a note
at the bottom and a new file is started.

This enables a user to leave evemu running in the background for a prolonged
time and recover the most recent recording after triggering a bug without
having to wade through three days of recordings.

https://bugs.freedesktop.org/show_bug.cgi?id=93752

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>