python: automatically rewind the fd at the end of the event sequence
Allows running through the event sequence more than once without any extra
handling. In modern recordings the description and the event file is in the
same file and there is no extra fd handling for the events. So simple code to
check events would look like this:

   d = evemu.Device("/path/to/file", create=False)
   for e in d.events():
   	check_for_something()

   for e in d.events():
   	check_for_something_else()

Simply rewinding the fd is sufficient here to avoid the caller having to keep
a copy of the events.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Daniel Martin <consume.noise@gmail.com>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>